1480: extract npm process as interface
to allow introducing another implementation
